Cleaned up the remaining packages.

+++ lib/zen/package/custom_fields/lib/custom_fields/controller/custom_field_types.rb
@@ -179,6 +179,8 @@ module CustomFields
         request.params['custom_field_type_ids'].each do |id|
           type = ::CustomFields::Model::CustomFieldType[id]
 
+          next if type.nil?
+
           begin
             type.destroy
           rescue => e
